drivers: ethernet: xmc4xxx: Fixes the checksum of bridged packets
Due to checksum offloading, the checksum field of a packet must be
zero before sending it to packet to the ethernet controller,
otherwise the computed checksum will be incorrect.

If the packet is bridged, it already has the correct checksum in
place, which will cause the automatically computed checksum to
incorrectly become zero.

In this patch we give bridged packets the special attention they
need.

Signed-off-by: Marcelo Roberto Jimenez <marcelo.jimenez@gmail.com>