drivers: ethernet: xmc4xxx: BUG Fix: dropping packets of certain lenghts on RX
The ethernet driver code was causing the silent dropping of packets with
length 125, 126 and 127. The root cause of the problem is that the DMA
engine performs an extra four bytes transfer corresponding to the Receive
Status Word. For those particular packet lengths, the previous code
incorrectly calculated the last fragment size.

The new code avoids that pitfall by decrementing the total frame size as
the fragments are being collected, and avoids creating and extra fragment
on the extracted network packet.

Signed-off-by: Marcelo Roberto Jimenez <marcelo.jimenez@gmail.com>